+++ backend/src/neuroagent/tools/generate_plot.py
@@ -1,6 +1,5 @@
 """Tool for generating plots with user-specified data and saving to storage."""
 
-import json
 import logging
 from typing import Any, ClassVar, Literal
 
@@ -74,6 +73,12 @@ class PlotMetadata(BaseMetadata):
     thread_id: str
 
 
+class PlotGeneratorToolOutput(BaseModel):
+    """Output class for the plot generator."""
+
+    storage_id: str
+
+
 class PlotGeneratorTool(BaseTool):
     """Tool that generates plots from user data and saves them to storage."""
 
@@ -91,10 +96,10 @@ class PlotGeneratorTool(BaseTool):
     • Histograms - For displaying distribution of numerical data
     • Line charts - For showing trends over a continuous range"""
 
-    input_schema: PlotInput
     metadata: PlotMetadata
+    input_schema: PlotInput
 
-    async def arun(self) -> str:
+    async def arun(self) -> PlotGeneratorToolOutput:
         """Generate plot and save to storage."""
         logger.info(f"Generating {self.input_schema.plot_type}")
 
@@ -176,10 +181,7 @@ class PlotGeneratorTool(BaseTool):
             thread_id=self.metadata.thread_id,
         )
 
-        return_dict = {
-            "storage_id": identifier,
-        }
-        return json.dumps(return_dict)
+        return PlotGeneratorToolOutput(storage_id=identifier)
 
     @classmethod
     async def is_online(cls) -> bool: