+++ backend/src/neuroagent/tools/get_morpho_tool.py
@@ -1,6 +1,5 @@
 """Get Morpho tool."""
 
-import json
 import logging
 from typing import Any, ClassVar
 
@@ -55,6 +54,12 @@ class KnowledgeGraphOutput(BaseModel):
     subject_age: str | None
 
 
+class GetMorphoToolOutput(BaseModel):
+    """Output schema for the Morpho tool."""
+
+    morphologies: list[KnowledgeGraphOutput]
+
+
 class GetMorphoTool(BaseTool):
     """Class defining the Get Morpho logic."""
 
@@ -82,10 +87,10 @@ class GetMorphoTool(BaseTool):
     • Access detailed morphological data
 
     Specify brain region and optional criteria to find relevant morphologies."""
-    input_schema: GetMorphoInput
     metadata: GetMorphoMetadata
+    input_schema: GetMorphoInput
 
-    async def arun(self) -> str:
+    async def arun(self) -> GetMorphoToolOutput:
         """From a brain region ID, extract morphologies.
 
         Returns
@@ -195,7 +200,7 @@ class GetMorphoTool(BaseTool):
         return entire_query
 
     @staticmethod
-    def _process_output(output: Any) -> str:
+    def _process_output(output: Any) -> GetMorphoToolOutput:
         """Process output to fit the KnowledgeGraphOutput pydantic class defined above.
 
         Parameters
@@ -231,10 +236,10 @@ class GetMorphoTool(BaseTool):
                     if "subjectAge" in res["_source"]
                     else None
                 ),
-            ).model_dump()
+            )
             for res in output["hits"]["hits"]
         ]
-        return json.dumps(formatted_output)
+        return GetMorphoToolOutput(morphologies=formatted_output)
 
     @classmethod
     async def is_online(