+++ backend/tests/tools/test_electrophys_tool.py
@@ -10,6 +10,7 @@ from neuroagent.tools import ElectrophysFeatureTool
 from neuroagent.tools.electrophys_tool import (
     CALCULATED_FEATURES,
     AmplitudeInput,
+    ElectrophysFeatureToolOutput,
     ElectrophysInput,
     ElectrophysMetadata,
 )
@@ -61,11 +62,10 @@ class TestElectrophysTool:
             ),
         )
         response = await tool.arun()
-        assert isinstance(response, str)
-        response = json.loads(response)
-        assert len(response["feature_dict"].keys()) == 1
+        assert isinstance(response, ElectrophysFeatureToolOutput)
+        assert len(response.feature_dict.keys()) == 1
         assert (
-            len(response["feature_dict"]["step_0"].keys())
+            len(response.feature_dict["step_0"].keys())
             == 2  # mean_frequency + 1 for stimulus current added manually
         )
 
@@ -114,11 +114,10 @@ class TestElectrophysTool:
             ),
         )
         response = await tool.arun()
-        assert isinstance(response, str)
-        response = json.loads(response)
-        assert len(response["feature_dict"].keys()) == 1
+        assert isinstance(response, ElectrophysFeatureToolOutput)
+        assert len(response.feature_dict.keys()) == 1
         assert (
-            len(response["feature_dict"]["step_0.25"].keys())
+            len(response.feature_dict["step_0.25"].keys())
             == 2  # mean_frequency + 1 for stimulus current added manually
         )
 
@@ -167,11 +166,10 @@ class TestElectrophysTool:
 
         # Without stimuli types and calculated features
         response = await tool.arun()
-        assert isinstance(response, str)
-        response = json.loads(response)
-        assert len(response["feature_dict"].keys()) == 1
+        assert isinstance(response, ElectrophysFeatureToolOutput)
+        assert len(response.feature_dict.keys()) == 1
         assert (
-            len(response["feature_dict"]["step_0"].keys())
+            len(response.feature_dict["step_0"].keys())
             == len(list(CALCULATED_FEATURES.__args__[0].__args__))
             + 1  # 1 for stimulus current added manually
         )