+++ CHANGELOG.md
@@ -12,6 +12,7 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 - Update system prompt.
 - Add current time and date in system prompt.
 ### Added
+- Every tool output schema to the OpenAPI json + swagger.
 - Header containing rate limit info.
 
 ## [v0.4.5] - 28.03.2025